requirements-dev.txt
# Development dependencies
aws-cdk-lib==2.8.0
constructs>=10.0.0,<11.0.0
boto3==1.18.21

# Development tools
pytest==6.2.4
moto==2.2.9
pylint==2.7.2
black==20.8b1
isort==5.8.0